Fix an inconsistency in the C5 text post-Metaverse
With the changes in 4c5024627e02a128eec75361ba368ca72d5d97b4, the C5 text post-Metaverse had an inconsistency, where the "Lifespan is now Infinity" part didn't match the "Lifespan is now infinite" the first line in the text of the Metaverse tab was changed to.

Change "Lifespan is now Infinity" to "infinite lifespan" in the C5 text post-Metaverse to resolve the inconsistency.